uk-train-chatbot/requirements.txt
anthropic>=0.36.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.1
requests>=2.32.3
pydantic>=2.8.2
apscheduler>=3.10.4
email-validator>=2.2.0
zeep>=4.2.1 
httpx>=0.27.0